% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IM.search.R
\name{IM.search}
\alias{IM.search}
\title{QTL search by IM}
\usage{
IM.search(
  marker,
  geno,
  y,
  method = "EM",
  type = "RI",
  D.matrix = NULL,
  ng = 2,
  cM = TRUE,
  speed = 1,
  crit = 10^-5,
  d.eff = FALSE,
  LRT.thre = TRUE,
  simu = 1000,
  alpha = 0.05,
  detect = TRUE,
  QTLdist = 15,
  plot.all = TRUE,
  plot.chr = TRUE,
  console = TRUE
)
}
\arguments{
\item{marker}{matrix. A k*2 matrix contains the marker information,
where the row dimension k is the number of markers in the chromosomes.
The first column labels the chromosomes where the markers are located,
and the second column labels the positions of QTLs (in morgan (M) or
centimorgan (cM)). Note that chromosome and position must be divided
in order.}

\item{geno}{matrix. A n*k matrix contains the k markers of the n
individuals. The marker genotypes of P1 homozygote (MM),
heterozygote (Mm) and P2 homozygote (mm) are coded as 2, 1 and 0,
respectively, and NA for missing value.}

\item{y}{vector. A vector with n elements that contains the phenotype
values of individuals.}

\item{method}{character. method="EM" means the interval mapping method
by Lander and Botstein (1989) is used in the analysis, while
method="REG" means  the approximate regression interval mapping method
by Haley and Knott (1992) is considered in the analysis.}

\item{type}{character. The population type of the dataset. Include
backcross (type="BC"), advanced intercross population (type="AI"), and
recombinant inbred population (type="RI").}

\item{D.matrix}{matrix. The design matrix of the IM model. If
D.matrix=NULL, the design matrix will be the constructed using the
Cockerham’s model. In BC population, it is a 2*1 matrix which is
0.5, -0.5 for additive effect. In RI or AI population, it is a 3*2 matrix
whose first column is 1, 0, -1 for additive effect and second column is
0.5, -0.5, 0.5 for dominant effect.}

\item{ng}{integer. The generation number of the population type. For
example, the BC1 population is type="BC" with ng=1; the AI F3
population is type="AI" with ng=3.}

\item{cM}{logical. Specify the unit of marker position. cM=TRUE for
centi-Morgan. Or cM=FALSE for Morgan.}

\item{speed}{numeric. The walking speed of the QTL search (in cM).}

\item{crit}{numeric. The convergent criterion of EM algorithm.
The E and M steps will be iterated until a convergent criterion
is satisfied.}

\item{d.eff}{logical. Specify whether the dominant effect will be
considered in the parameter estimation or not for AI or RI population.}

\item{LRT.thre}{logical or numeric. If being TRUE, the LRT threshold
will be computed based on the Gaussian stochastic process
(Kao and Ho 2012). Or users can input a numerical value as the LRT
threshold to assessing the significance of QTL detection.}

\item{simu}{integer. To decide how many simulation samples will be used
to compute the LRT threshold using the Gaussian process.}

\item{alpha}{numeric. The type I error rate for the LRT threshold.}

\item{detect}{logical. Whether the significant QTL whose LRT statistic
is larger than the LRT threshold will be shown in the output dataset or
not.}

\item{QTLdist}{numeric. The minimum distance (cM) among different
linked significant QTL.}

\item{plot.all}{logical. If being TRUE, output the profile of LRT
statistics for the genome in one figure.}

\item{plot.chr}{logical. If being TRUE, output the profile of LRT
statistics for the chromosomes.}

\item{console}{logical. To decide whether the process of algorithm will
be shown in the R console or not.}
}
\value{
\item{effect}{The estimated effects and LRT statistics of all positions.}
\item{LRT.threshold}{The LRT threshold value computed for the data using the
Gaussian stochastic process (Kuo 2011; Kao and Ho 2012).}
\item{detect.QTL}{The positions, effects and LRT statistics of the detected
QTL significant using the obtained LRT threshold value.}

Graphical outputs including LOD value and effect of each position.
}
\description{
Expectation-maximization algorithm for QTL interval mapping to search
the possible position of QTL in all chromosome.
}
\examples{
# load the example data
load(system.file("extdata", "exampledata.RDATA", package = "QTLEMM"))

# run and result
result <- IM.search(marker, geno, y, type = "RI", ng = 2, speed = 7.5, crit = 10^-3, LRT.thre = 10)
result$detect.QTL
}
\references{
KAO, C.-H. and Z.-B. ZENG 1997 General formulas for obtaining the maximum
likelihood estimates and the asymptotic variance-covariance matrix in QTL
mapping when using the EM algorithm. Biometrics 53, 653-665.

KAO, C.-H., Z.-B. ZENG and R. D. TEASDALE 1999 Multiple interval mapping
for Quantitative Trait Loci. Genetics 152: 1203-1216.

KAO, C.-H. and H.-A. Ho 2012 A score-statistic approach for determining
threshold values in QTL mapping. Frontiers in Bioscience. E4, 2670-2682.
}
\seealso{
\code{\link[QTLEMM]{EM.MIM}}
\code{\link[QTLEMM]{IM.search2}}
\code{\link[QTLEMM]{LRTthre}}
}
